Added onDrop for each component.
