+++ ts/Dashboards/Components/HTMLComponent.ts
@@ -23,11 +23,12 @@
  * */
 
 import type Cell from '../Layout/Cell.js';
+import type Row from '../Layout/Row.js';
+import type SidebarPopup from '../EditMode/SidebarPopup.js';
 
 import AST from '../../Core/Renderer/HTML/AST.js';
 import Component from './Component.js';
 import U from '../../Core/Utilities.js';
-
 const {
     merge,
     diffObjects
@@ -257,6 +258,21 @@ class HTMLComponent extends Component {
     public async update(options: Partial<HTMLComponent.HTMLComponentOptions>): Promise<void> {
         await super.update(options);
         this.emit({ type: 'afterUpdate' });
+    }
+
+    public onDrop(sidebar: SidebarPopup, dropContext: Cell|Row): void|Cell {
+        if (sidebar && dropContext) {
+            return sidebar.onDropNewComponent(dropContext, {
+                cell: '',
+                type: 'HTML',
+                elements: [{
+                    tagName: 'img',
+                    attributes: {
+                        src: 'https://www.highcharts.com/samples/graphics/stock-dark.svg'
+                    }
+                }]
+            });
+        }
     }
 
     /**
